Handle piping to and from tag
Previously if you piped the output of tag to a file, it would not
correctly be reformatted by ag or rg. This meant you would end up with
output that wasn't parseable on a single line, and had color codes, even
though they wouldn't particularly be handled. Now we check to see if
either stdin or stdout is not a terminal, and we pass through the
command directly, without any processing.